split m_summary into several files
