+++ packages/devextreme/js/__internal/grids/data_grid/summary/const.ts
@@ -1,0 +1,11 @@
+export const DATAGRID_SUMMARY_ITEM_CLASS = 'dx-datagrid-summary-item';
+export const DATAGRID_TOTAL_FOOTER_CLASS = 'dx-datagrid-total-footer';
+export const DATAGRID_TEXT_CONTENT_CLASS = 'dx-datagrid-text-content';
+export const DATAGRID_GROUP_FOOTER_CLASS = 'dx-datagrid-group-footer';
+export const DATAGRID_GROUP_TEXT_CONTENT_CLASS = 'dx-datagrid-group-text-content';
+export const DATAGRID_NOWRAP_CLASS = 'dx-datagrid-nowrap';
+export const DATAGRID_FOOTER_ROW_CLASS = 'dx-footer-row';
+export const DATAGRID_CELL_DISABLED = 'dx-cell-focus-disabled';
+
+export const DATAGRID_GROUP_FOOTER_ROW_TYPE = 'groupFooter';
+export const DATAGRID_TOTAL_FOOTER_ROW_TYPE = 'totalFooter';
